renamed cosmwasm folder and parsers names
