+++ cosmwasm/query_parser.go
@@ -1,0 +1,1 @@
+package cosmwasm
