--- cosmwasm_plugins/query_parser.go
@@ -1,1 +1,0 @@
-package cosmwasm_plugins
